+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.17.0-0.7.3 (upcoming)
 
+* [PLT-1548] [GKE] Activar Workload Identity
 * [PLT-2389] Fix `cloud-provisioner` image reference during cluster creation
 
 ## 0.17.0-0.7.2 (2025-06-19)